+++ feature-governance-impl/src/main/java/io/novafoundation/nova/feature_governance_impl/domain/summary/ReferendaSummaryInteractor.kt
@@ -1,10 +1,14 @@
 package io.novafoundation.nova.feature_governance_impl.domain.summary
 
 import io.novafoundation.nova.feature_governance_api.data.network.blockhain.model.ReferendumId
-import io.novafoundation.nova.feature_governance_api.domain.referendum.summary.ReferendaSummaryInteractor
 import io.novafoundation.nova.feature_governance_impl.data.GovernanceSharedState
 import io.novafoundation.nova.runtime.state.selectedOption
 import kotlinx.coroutines.CoroutineScope
+
+interface ReferendaSummaryInteractor {
+
+    suspend fun getReferendaSummaries(ids: List<ReferendumId>, coroutineScope: CoroutineScope): Map<ReferendumId, String>
+}
 
 class RealReferendaSummaryInteractor(
     private val governanceSharedState: GovernanceSharedState,
